requirements.txt
django
django-rest-framework
requests
jinja2
psycopg2